Flask/requirements.txt
Flask==2.3.2
flask-cors==3.1.1
pyodbc==5.1.0
ldap3==2.9.1
requests==2.31.0
pywinrm==0.4.3
APScheduler